+++ components/upload/demo/custom-request.tsx
@@ -1,11 +1,9 @@
+/* eslint-disable @typescript-eslint/no-use-before-define */
 /* eslint-disable no-await-in-loop */
 import * as React from 'react';
 import type { UploadProps } from 'antd';
-import { Button, Upload, Divider } from 'antd';
+import { Button, Upload, Divider, Form, InputNumber, Switch, Input } from 'antd';
 import { UploadOutlined } from '@ant-design/icons';
-import type { FetchWithRetryOptions } from './custom-request/useFetch';
-import { fetchWithRetry } from './custom-request/useFetch';
-import useUploadConfig from './custom-request/useUploadConfig';
 
 const App: React.FC = () => {
   const [config, contentHelper] = useUploadConfig({
@@ -86,3 +84,87 @@ const App: React.FC = () => {
 };
 
 export default App;
+
+type RetryOptions = {
+  count: number;
+  delay: number;
+};
+
+type FetchWithRetryOptions = {
+  retry?: number | RetryOptions;
+} & RequestInit;
+
+async function fetchWithRetry(url: string, options?: FetchWithRetryOptions): Promise<any> {
+  const { retry, ...fetchOptions } = options || {};
+
+  const retries = typeof retry === 'number' ? retry : retry?.count || 3;
+  const retryDelay = (typeof retry === 'object' && retry.delay) || 3000;
+
+  let error = null;
+
+  for (let i = 0; i <= retries; i++) {
+    try {
+      // eslint-disable-next-line compat/compat
+      const response = await fetch(url, fetchOptions);
+      const data = await response.json();
+      return { data, response };
+    } catch (e) {
+      error = e;
+    }
+
+    if (i < retries) {
+      await new Promise((resolve) => {
+        setTimeout(resolve, retryDelay);
+      });
+    }
+  }
+
+  throw error;
+}
+
+function useUploadConfig(config: any) {
+  const [state, setState] = React.useState(config);
+  const [form] = Form.useForm();
+
+  const inputChunkSize = Form.useWatch('chunkSize', form);
+  const [size, sizeM, sizeK] = React.useMemo(() => {
+    const _size = inputChunkSize || 0;
+    const M = Number.prototype.toFixed.call(_size / 1024 / 1024, 2);
+    const K = Number.prototype.toFixed.call(_size / 1024, 2);
+    return [_size, M, K];
+  }, [inputChunkSize]);
+
+  const contentHelper = (
+    <Form form={form} initialValues={state} onValuesChange={(_, allValues) => setState(allValues)}>
+      <Form.Item label="Url" name="url">
+        <Input />
+      </Form.Item>
+      <Form.Item label="IsChunk" name="isChunk" valuePropName='checked'>
+        <Switch />
+      </Form.Item>
+      <Form.Item
+        label="ChunkSize(byte)"
+        name="chunkSize"
+        validateTrigger="blur"
+        help={`${size}byte ≈ ${sizeK}kb ≈ ${sizeM}mb`}
+      >
+        <InputNumber
+          disabled={!state.isChunk}
+          min={10 * 1024}
+          max={10 * 1024 * 1024}
+          step={10 * 1024}
+        />
+      </Form.Item>
+      <Form.Item name="retry">
+        <Form.Item label="retry" name={['retry', 'count']}>
+          <InputNumber min={1} max={10} />
+        </Form.Item>
+        <Form.Item label="retry(ms)" name={['retry', 'delay']}>
+          <InputNumber min={100} max={30 * 1000} step={200} />
+        </Form.Item>
+      </Form.Item>
+    </Form>
+  );
+
+  return [state, contentHelper] as const;
+}
